Fixed Everything
Alright, Amjad, don't delete code if it's working.  I had to go back and
put everything back in.  If it's not hurting anything, don't touch it.
just focus on the admin page and the index if you want to.  Everything
should be working now.  and yes, there's a reason for having
"session_start();" and including the datalogin on other pages.  If you
want to talk about it later that's fine, but for now just don't touch
it.