edited to query tblStudent instead of tblAdmin
Changed the query to tblStudents instead of the admin table, and the if else statement so that if the superadmin table contains 'Y' then they are set as true for super admin. I noticed when you login you still have access to add courses even when your not an admin so im not sure if there is a problem in the login query or if its just not being checked yet.